+++ dsl/src/main/scala/org/http4s/dsl/Http4sDsl.scala
@@ -28,6 +28,7 @@ trait Http4sDsl2[F[_], G[_]] extends RequestDsl with Statuses with Responses[F, 
   val :? : impl.:?.type = impl.:?
   val ~ : impl.~.type = impl.~
   val -> : impl.->.type = impl.->
+  val ->> : impl.->>.type = impl.->>
   val /: : impl./:.type = impl./:
   val +& : impl.+&.type = impl.+&
 
@@ -54,8 +55,7 @@ trait Http4sDsl[F[_]] extends Http4sDsl2[F, F] {
 }
 
 object Http4sDsl {
-  // Does not return Http4sDslBinCompat for bincompat reasons. ¯\_(ツ)_/¯
-  def apply[F[_]]: Http4sDsl[F] with RequestDslBinCompat = new Http4sDslBinCompat[F] {}
+  def apply[F[_]]: Http4sDsl[F] = new Http4sDsl[F] {}
 
   final class MethodOps(val method: Method) extends AnyVal {
     def |(another: Method) = new MethodConcat(Set(method, another))
@@ -65,5 +65,3 @@ object Http4sDsl {
     def |(another: Method) = new MethodConcat(methods.methods + another)
   }
 }
-
-trait Http4sDslBinCompat[F[_]] extends Http4sDsl[F] with RequestDslBinCompat